[IMP] project_timesheet_holidays: add hook to decide when to delete analytic lines linked to the time off: For automated processes that massively approve leaves this deleting may result in unwated deletions. For example, approving public holidays for the next year.

+++ addons/project_timesheet_holidays/models/hr_holidays.py
@@ -69,10 +69,11 @@ class Holidays(models.Model):
             l.holiday_status_id.timesheet_project_id.sudo().company_id == (l.holiday_status_id.company_id or self.env.company))
 
         # Unlink previous timesheets do avoid doublon (shouldn't happen on the interface but meh)
-        old_timesheets = holidays.sudo().timesheet_ids
-        if old_timesheets:
-            old_timesheets.holiday_id = False
-            old_timesheets.unlink()
+        if not self.env.context.get('skip_unlink_timesheets'):
+            old_timesheets = holidays.sudo().timesheet_ids
+            if old_timesheets:
+                old_timesheets.holiday_id = False
+                old_timesheets.unlink()
 
         # create the timesheet on the vacation project
         holidays._timesheet_create_lines()
